split thunks and slice into separate files
